requirements.txt
black
tree-sitter
openai
transformers
torch
tiktoken
google-generativeai